Perform an integrity check against the opened database using "PRAGMA integrity_check" command, if there is any error found then we attempt to remove the corrupt database and recreate a new one and we prevent GAI from using a corrupt database.
This is to fix the problem that the `hits` table in the SQLite database is corrupt for some reason which ends up with enumerous repeated hits stored in that table, and we can't count them or delete any of them, but SQLite is a strong database which allows us to fetch data from the `hits` table even if it's corrupt. Hence, the dispatch thread is stuck in the while loop forever as it can always fetch hits from the corrupt table but it cannot delete the fetched hits from it which results in an ever-increasing memory allocation situation (we request a page via the POST protocol for each hit which allocates memory, and we cannot deallocate the memory since we are stuck in the loop).