Fix dispatcher losing hits on post failure
Previously the dispatcher would remove a hit before a post request was sent, and
add the hit back to the datastore if the response failed.

This meant that if no response was received the hit was lost, which was prevalent
at shutdown or when no network was available.

Hits are no longer removed from the datastore when a post request is sent, they
are now removed when a successful response is received.

Added the below tests to cover the above changes.
- DispatcherTest.hits_are_removed_on_successful_response
- DispatcherTest.hits_are_not_removed_on_failed_response
- DispatcherTest.hits_are_not_removed_on_no_response

Tidied up some other tests.